Delete Brian Mallari - Some Exploratory Analysis - UK Train Rides (MOCK DATA).xlsx
I don't know how to move a file from one repository to another, so I just uploaded this file again into a new repository, and I'll be deleting this copy in this repository. Any future changes to the file can be applied to the new respository, and any new files related to data analytics can go into that repsoitory.