Update index.md
I've updated the links to the Python projects under "Guided Projects" so that they link to the repositories for those projects instead of to the Jupyter Notebook files. I corrected the text for the correlations project. 

I didn't use the project name in the text because I'd have ended up reapting the word "Python" twice. Also, I don't plan on changing the titles of the project or the repository to reflect the topic of the underlying data (movies) like in the case of the other projects because that'd be a lot of work, and I'm not sure if the reward is worth all the effort, ya know?